pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "elmesh"
version = "1.0.0"
description = "Create 3D mesh objects from elevation data"
readme = "README.md"
authors = [{ name = "Paul Ockenfuß", email = "paul.ockenfuss@physik.uni-muenchen.de" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Science/Research",
]
keywords = ["mesh", "terrain", "elevation", "dem"]
dependencies = [
    "numpy-stl"
]
requires-python = ">=3.8"

# [project.optional-dependencies]
# extra = ["pyyaml"]

[project.urls]
Homepage = "https://github.com/Ockenfuss/elmesh"

# [project.scripts]
# inlog-flowchart = "inlog.flowchart:main"